+++ app/email.py
@@ -1,6 +1,5 @@
 from flask import render_template, url_for, redirect, flash
 from flask_mail import Message
-from flask_login import login_user
 from app import app, db, mail
 from app.models import User
 from app.token import generate_confirmation_token
@@ -32,8 +31,7 @@ def send_contact_email(form):
 
 def send_reset_email(user):
     token = user.reset_token()
-    #message = Message('[Knewbie] Password Reset Request', sender='resetpassword@knewbie.com', recipients=[user.email])
-    send_email('Password Reset Request', 
+    send_email('[Knewbie] Password Reset Request', 
                sender=app.config['ADMINS'][0],
                recipients=[user.email],
                text_body=render_template('email/reset.txt', name=user.firstName, token=token),
@@ -42,7 +40,6 @@ def send_reset_email(user):
 
 def send_deactivate_email(user):
     token = user.reset_token()
-    #message = Message('[Knewbie] Deactivate Account Request', sender='deactivate@knewbie.com', recipients=[user.email])
     send_email('[Knewbie] Deactivate Account Request', 
                sender=app.config['ADMINS'][0],
                recipients=[user.email],
@@ -55,28 +52,13 @@ def get_confirm_url(user):
     confirm_url = url_for('confirm_email', token=token, _external=True)
     return confirm_url
 
-def register(form, role):
-    user = User(firstName=form.firstName.data, lastName=form.lastName.data, email=form.email.data, urole=role, confirmed=False)
-    user.set_password(form.password.data)
-    if role == 'student':
-        set_knewbie_id(user)
-    db.session.add(user)
-    db.session.commit()
-    confirm_url = get_confirm_url(user)
-    send_conf_email(user, confirm_url)
-        
-    login_user(user)
-
-    flash('A confirmation email has been sent via email.', 'success')
-    #flash('Congratulations, you are now a registered user!')
-    return redirect(url_for('unconfirmed'))
-
 def resend_conf(user):
     confirm_url = get_confirm_url(user)
     send_conf_email(user, confirm_url)
     flash('A new confirmation email has been sent.', 'success')
 
-
+###########################
+# To move to Unit Testing #
 def add_user():
     remove_users()
     if User.query.all(): return
