+++ app/forms.py
@@ -1,9 +1,9 @@
+from flask_login import current_user
 from flask_wtf import FlaskForm
+from flask_wtf.file import FileAllowed
 from wtforms import StringField, PasswordField, BooleanField, SubmitField, TextAreaField, MultipleFileField, SelectField, FileField
 from wtforms.validators import ValidationError, DataRequired, Email, EqualTo, Length
-from flask_login import current_user
 
-from flask_wtf.file import FileAllowed
 from app.models import User, Topic
 
 class LoginForm(FlaskForm):
@@ -24,11 +24,6 @@ class RegistrationForm(FlaskForm):
     password2 = PasswordField(
         'Repeat Password', validators=[DataRequired(), EqualTo('password')])
     submit = SubmitField('Register')
-
-    #def validate_username(self, username):
-    #    user = User.query.filter_by(username=username.data).first()
-    #    if user is not None:
-    #        raise ValidationError('Please use a different username.')
 
     def validate_email(self, email):
         user = User.query.filter_by(email=email.data).first()
