+++ app/group.py
@@ -1,6 +1,7 @@
+from flask_login import current_user
 from app import db
 from app.models import User, Group
-from flask_login import current_user
+from app.profile import set_code
 
 def validate_group_link(groupID):
     return Group.query.filter_by(id=groupID).filter(Group.users.any(id=current_user.id)).first_or_404()
@@ -15,12 +16,20 @@ def get_sorted_students(groupID):
     return User.query.filter_by(urole='student').\
         filter(User.groups.any(id=groupID)).order_by(User.curr_theta.desc()).all()
 
+def set_class_code(group):
+    code = set_code(6)
+    while Group.query.filter_by(classCode=code).first():
+        code = set_code(6)
+    group.classCode = code
+    return group
+
 def add_group(name):
     group = Group(name=name)
     set_class_code(group)
-    group.users.append(current_user)
+    add_user(group, current_user)
     db.session.add(group)
     db.session.commit()
+    return group
 
 def add_user(group, user):
     if user in group.users: return #prevent duplicates
